Add support for including additional CSS files
We can use this to include CSS we need to make sure is always present
in the critical CSS.

The support is a bit weird. Because our CSS files have hashes in their
names after an asset build, we can't easily specify a fixed URL to
always include. Instead, we allow the user to specify a pattern, which
will be matched against all the hrefs we gather from the page when
looking for `<link>` tags. (Technically, we're searching for `<style>`
and `<link>` tags, but we'll skip tags without an `href` attribute,
which leaves out `<style>` tags.) We download all the URLs which match
that pattern and include their contents in the output. Note that we're
probably going to dump a substantial amount of duplicate CSS into the
resulting output file. We're okay with that, because the `postcss`
postprocessing that we do should eliminate it pretty completely (we've
got it configured to combine duplicated selectors and remove
duplicated properties from the results).