Check regex in a string properly
Created a new function to check if the string contains a regex.
First it checks if it has more than 2 forward slashes and gets
the pattern and flags. Then we finally check if it's a regex.
If it's not we just return it as a string.

Then the final array is returned as mix of strings and regexes.