CSS Gather: Fix flakiness
This took quite some time to track down. We were fetching the critical
css for all pages in the url list asynchronously. This was fine, but
we were also outputting to stdout asynchronously. This meant that the
css that we ran through our postprocessing scripts later would always
receive the css in a different order.

Now, we still fetch the css asynchronously so it's quick, but we
always output to stdout in the same order as the input urls array.